pyproject.toml
[tool.poetry]
name = "arxbot"
version = "0.1.0"
description = "Yet Another GW2 Discord Bot"
authors = ["harasho"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = "^2.3.2"
python-dotenv = "^1.0.1"
peewee = "^3.17.5"
requests = "^2.32.2"
aiohttp = "^3.9.5"
tabulate = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
